% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mlb_2008}
\alias{mlb_2008}
\title{A dataset containing individual offensive statistics from the 2008 Major League Baseball season.}
\format{
A data frame with 337 rows and 134 variables:
\describe{
  \item{player}{name}
  \item{record_id}{unknown}
  ...
}
}
\source{
\url{http://www.exploredata.net/Downloads/Baseball-Data-Set}
}
\usage{
mlb_2008
}
\description{
A dataset containing individual offensive statistics from the 2008 Major League Baseball season.
}
\keyword{datasets}
