% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spellman}
\alias{spellman}
\title{A dataset based on the CDC15 yeast gene expression data set of Spellman et al. curated to deal with issues of missing data.}
\format{
A data frame with 4381 rows and 24 variables:
\describe{
  \item{Time}{Time}
  \item{X40}{A gene}
  ...
}
}
\source{
\url{http://www.exploredata.net/Downloads/Gene-Expression-Data-Set}
}
\usage{
spellman
}
\description{
A dataset based on the CDC15 yeast gene expression data set of Spellman et al. curated to deal with issues of missing data.
}
\keyword{datasets}
