% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymine.R
\name{tidy_mine_output}
\alias{tidy_mine_output}
\title{Tidy the output of minerva::mine}
\usage{
tidy_mine_output(mine_output, predictor_variable_names)
}
\arguments{
\item{mine_output}{The output from running minerva::mine}

\item{predictor_variable_names}{A vector of the variable names used as $x$ in minerva::mine.}
}
\value{
A tidy tibble of MINE statistics.
}
\description{
Tidy the output of minerva::mine
}
\examples{
x <- mlb_2008[,-c(1:3)]
excl <- which(diag(var(x)) < 1e-5) # exclude variables with low variance
x <- x[,-excl]
y <- dat$SALARY
minerva_result <- minerva::mine(x = x, y = y)
tidy_mine_output(minerva_result, x)
}
