% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidymine.R
\name{tidymine}
\alias{tidymine}
\title{Calculate MINE statistics.}
\usage{
tidymine(
  dataset,
  outcome,
  predictors = -{     {         outcome     } },
  alpha = 0.6,
  c = 15,
  cores = getOption("mc.cores", 1L),
  low_variance_threshold = 1e-05,
  eps = NULL,
  est = "mic_approx",
  na.rm = FALSE,
  use = "pairwise.complete.obs",
  normalisation = FALSE,
  ...
)
}
\arguments{
\item{dataset}{The dataset containing the outcome and predictors.}

\item{outcome}{The variable of interest.}

\item{predictors}{The variables to measure the predictive power of. Defaults to every variable apart from the outcome.}

\item{cores}{Number of cores to use when executing the chains in parallel.
This defaults to 1 (i.e. not performing parallel computing).
Running `options(mc.cores = parallel::detectCores())` will set this be the maximum number of processors as the hardware and RAM allow (up to the number of chains).}

\item{low_variance_threshold}{A parameter to remove predictors with low variance. MINE statistics cannot be calculated for variables With low variance. This is fine because a low variance predictor cannot have high predictive power.}

\item{est}{Default value is "mic_approx".
With est="mic_approx" the original MINE statis-tics will be computed,
with est="mic_e" the equicharacteristic matrix is is evaluated and the mic() and tic() methods
will return MIC_e and TIC_e values respectively.}

\item{na.rm}{Logical to remove NA values.}

\item{use}{Method to deal with NA values.}

\item{...}{}
}
\value{
MINE statistics of \code{predictors} for predicting \code{outcome}.
}
\description{
Calculate MINE statistics.
}
\examples{
spellman \%>\%
  tidymine(x40)
who \%>\%
  tidymine(lung_cancer_deaths_per_100_000_women)
}
