% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{who}
\alias{who}
\title{A dataset of compiled social, economic, health, and political indicators using data from the World Health Organization and partner organizations.}
\format{
A data frame with 202 rows and 358 variables:
\describe{
  \item{Country}{Country name}
  \item{CountryID}{Country ID}
  ...
}
}
\source{
\url{http://www.exploredata.net/Downloads/WHO-Data-Set}
}
\usage{
who
}
\description{
A dataset of compiled social, economic, health, and political indicators using data from the World Health Organization and partner organizations.
}
\keyword{datasets}
